+++ service/database/is-banned.go
@@ -1,8 +1,6 @@
 package database
 
-
-
-func (db *appdbimpl) IsBanned(myId uint64, idProfile uint64) (bool, error) {
+func (db *appdbimpl) IsBanned(myId int, idProfile int) (bool, error) {
 	row := db.c.QueryRow(`SELECT * FROM Banned WHERE banned=? AND whoBan=?`, idProfile, myId)
 	text := ""
 	err := row.Scan(&text)
